+++ actionpack/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Rails 4.0.0 (unreleased) ##
 
+*   Allow value to be set on date_select tag helper.
+
+    *Colin Burn-Murdoch*
+
 *   Fixed json params parsing regression for non-object JSON content.
 
     *Dylan Smith*